+++ app/scripts/TiledPixiTrack.js
@@ -518,7 +518,7 @@ class TiledPixiTrack extends PixiTrack {
     return this.scale.maxRawValue;
   }
 
-  initTile(/* tile */) {
+  initTile(tile) {
     // create the tile
     // should be overwritten by child classes
     this.scale.minRawValue = this.continuousScaling