+++ app/scripts/configs/tracks-info.js
@@ -309,6 +309,7 @@ export const TRACKS_INFO = [
       'axisMargin',
       'lineStrokeWidth',
       'lineStrokeColor',
+      'nanAsZero',
       'valueScaling',
       'valueScaleMin',
       'valueScaleMax',