+++ karma.conf.js
@@ -46,6 +46,7 @@ module.exports = (config) => {
       'test/HiGlassComponentCreationTests.js',
       'test/HiGlassComponent/*.js',
       'test/Horizontal1DTrackTests.js',
+      'test/HorizontalLineTrackTests.js',
       'test/HorizontalHeatmapTests.js',
       'test/HorizontalMultivecTests.js',
       'test/LabelTests.js',